# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: CineWrapped API
  version: 1.0.0
  summary: Versioned REST contract for the CineWrapped MVP
  description: >-
    Product API consumed by CineWrapped mobile and admin clients. Supabase Auth
    issues identity sessions; this API owns application authorization and data.
  contact:
    name: CineWrapped Engineering
servers:
  - url: http://localhost:4000/api/v1
    description: Local development
  - url: https://api.staging.cinewrapped.example/api/v1
    description: Staging
  - url: https://api.cinewrapped.example/api/v1
    description: Production
security:
  - bearerAuth: []
...
paths:
...
  /users/me:
    get:
      tags: [Users]
      operationId: getCurrentUser
      x-response-dto: CurrentUser
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    patch:
      tags: [Users]
      operationId: updateCurrentUser
      requestBody: { $ref: '#/components/requestBodies/UpdateProfile' }
      x-response-dto: CurrentUser
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
    delete:
      tags: [Users]
      operationId: requestAccountDeletion
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody: { $ref: '#/components/requestBodies/DeleteAccount' }
      x-response-dto: DeletionRequest
      responses:
        '202': { $ref: '#/components/responses/Accepted' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /users/me/preferences:
    get:
      tags: [Users]
      operationId: getPreferences
      x-response-dto: UserPreferences
      responses: { '200': { $ref: '#/components/responses/Ok' } }
    patch:
      tags: [Users]
      operationId: updatePreferences
      requestBody: { $ref: '#/components/requestBodies/UpdatePreferences' }
      x-response-dto: UserPreferences
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /users/me/privacy:
    get:
      tags: [Users]
      operationId: getPrivacySettings
      x-response-dto: PrivacySettings
      responses: { '200': { $ref: '#/components/responses/Ok' } }
    patch:
      tags: [Users]
      operationId: updatePrivacySettings
      requestBody: { $ref: '#/components/requestBodies/UpdatePrivacy' }
      x-response-dto: PrivacySettings
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /users/me/onboarding:
    get:
      tags: [Onboarding]
      operationId: getOnboardingState
      x-response-dto: OnboardingState
      responses: { '200': { $ref: '#/components/responses/Ok' } }
    patch:
      tags: [Onboarding]
      operationId: updateOnboardingState
      requestBody: { $ref: '#/components/requestBodies/UpdateOnboarding' }
      x-response-dto: OnboardingState
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /users/me/onboarding/complete:
    post:
      tags: [Onboarding]
      operationId: completeOnboarding
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody: { $ref: '#/components/requestBodies/CompleteOnboarding' }
      x-response-dto: CurrentUser
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
...
  /recommendations/taste-profile:
    get:
      tags: [Recommendations]
      operationId: getTasteProfile
      x-response-dto: TasteProfile
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /recommendations/refresh:
    post:
      tags: [Recommendations]
      operationId: refreshRecommendations
      x-response-dto: RecommendationRefreshSummary
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /recommendations:
    get:
      tags: [Recommendations]
      operationId: listRecommendations
      parameters:
        - { name: type, in: query, schema: { $ref: '#/components/schemas/RecommendationType' } }
        - { name: runtimeMax, in: query, schema: { type: integer, minimum: 1, maximum: 1440 } }
        - { $ref: '#/components/parameters/Cursor' }
        - { $ref: '#/components/parameters/Limit' }
      x-response-dto: RecommendationPage
      responses:
        {
          '200': { $ref: '#/components/responses/Collection' },
          '403': { $ref: '#/components/responses/Forbidden' },
        }
  /recommendations/{recommendationId}/feedback:
    post:
      tags: [Recommendations]
      operationId: createRecommendationFeedback
      parameters:
        [
          { $ref: '#/components/parameters/RecommendationId' },
          { $ref: '#/components/parameters/IdempotencyKey' },
        ]
      requestBody: { $ref: '#/components/requestBodies/RecommendationFeedback' }
      x-response-dto: Recommendation
      responses:
        {
          '200': { $ref: '#/components/responses/Ok' },
          '404': { $ref: '#/components/responses/NotFound' },
          '409': { $ref: '#/components/responses/Conflict' },
        }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase-issued access token verified by the CineWrapped API.
  parameters:
    Cursor:
      {
        name: cursor,
        in: query,
        schema: { type: string, minLength: 8, maxLength: 2048 },
        description: Opaque cursor bound to route and filters.,
      }
    Limit:
      { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 50, default: 20 } }
    IdempotencyKey:
      {
        name: Idempotency-Key,
        in: header,
        required: true,
        schema: { type: string, minLength: 16, maxLength: 128, pattern: '^[\x21-\x7E]+$' },
      }
...
    RecommendationId:
      { name: recommendationId, in: path, required: true, schema: { type: string, format: uuid } }
...
  requestBodies:
...
    UpdateProfile:
      {
        required: true,
        content:
          { application/json: { schema: { $ref: '#/components/schemas/UpdateProfileRequest' } } },
      }
    UpdatePreferences:
      {
        required: true,
        content:
          {
            application/json: { schema: { $ref: '#/components/schemas/UpdatePreferencesRequest' } },
          },
      }
    UpdatePrivacy:
      {
        required: true,
        content:
          { application/json: { schema: { $ref: '#/components/schemas/UpdatePrivacyRequest' } } },
      }
    UpdateOnboarding:
      {
        required: true,
        content:
          {
            application/json: { schema: { $ref: '#/components/schemas/UpdateOnboardingRequest' } },
          },
      }
    CompleteOnboarding:
      {
        required: true,
        content:
          {
            application/json:
              { schema: { $ref: '#/components/schemas/CompleteOnboardingRequest' } },
          },
      }
...
    DeleteAccount:
      {
        required: true,
        content:
          { application/json: { schema: { $ref: '#/components/schemas/DeleteAccountRequest' } } },
      }
...
    RecommendationFeedback:
      {
        required: true,
        content:
          {
            application/json:
              { schema: { $ref: '#/components/schemas/RecommendationFeedbackRequest' } },
          },
      }
...
  responses:
    Ok:
      {
        description: Successful response.,
        headers: { X-Request-ID: { $ref: '#/components/headers/RequestId' } },
        content: { application/json: { schema: { $ref: '#/components/schemas/SuccessEnvelope' } } },
      }
...
    Accepted:
      {
        description: Request accepted for asynchronous processing.,
        headers: { X-Request-ID: { $ref: '#/components/headers/RequestId' } },
        content: { application/json: { schema: { $ref: '#/components/schemas/SuccessEnvelope' } } },
      }
    Collection:
      {
        description: Cursor-paginated collection.,
        headers: { X-Request-ID: { $ref: '#/components/headers/RequestId' } },
        content:
          { application/json: { schema: { $ref: '#/components/schemas/CollectionEnvelope' } } },
      }
...
    Unauthorized:
      {
        description: Authentication required or invalid.,
        content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } },
      }
    Forbidden:
      {
        description: Actor is not permitted.,
        content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } },
      }
    NotFound:
      {
        description: Resource absent or concealed.,
        content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } },
      }
    Conflict:
      {
        description: 'State, version, uniqueness, or idempotency conflict.',
        content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } },
      }
...
    ValidationError:
      {
        description: Domain validation failed.,
        content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } },
      }
    RateLimited:
      {
        description: Rate limit exceeded.,
        headers: { Retry-After: { schema: { type: integer } } },
        content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } },
      }
...
  headers:
    RequestId:
      {
        description: Correlation identifier.,
        schema: { type: string, minLength: 8, maxLength: 100 },
      }
  schemas:
...
    Visibility: { type: string, enum: [PUBLIC, FRIENDS, PRIVATE, CLUB_ONLY] }
...
    RecommendationType:
      {
        type: string,
        enum:
          [
            PERSONALIZED,
            TRENDING,
            FRIEND_BASED,
            MOOD_BASED,
            SIMILAR_MEDIA,
            HIDDEN_GEM,
            CONTINUE_WATCHING,
            BECAUSE_YOU_WATCHED,
          ],
      }
    RequestMeta:
      type: object
      additionalProperties: false
      required: [requestId]
      properties: { requestId: { type: string } }
    PageMeta:
      type: object
      additionalProperties: false
      required: [nextCursor, hasMore, limit]
      properties:
        nextCursor: { type: [string, 'null'] }
        hasMore: { type: boolean }
        limit: { type: integer, minimum: 1, maximum: 50 }
    SuccessEnvelope:
      type: object
      additionalProperties: false
      required: [success, data, meta]
      properties:
        success: { const: true }
        data: { oneOf: [{ type: object }, { type: array, items: {} }, { type: 'null' }] }
        meta: { $ref: '#/components/schemas/RequestMeta' }
    CollectionEnvelope:
      type: object
      additionalProperties: false
      required: [success, data, meta]
      properties:
        success: { const: true }
        data: { type: array, items: { type: object } }
        meta:
          allOf:
            - { $ref: '#/components/schemas/RequestMeta' }
            - type: object
              required: [page]
              properties: { page: { $ref: '#/components/schemas/PageMeta' } }
    ErrorObject:
      type: object
      additionalProperties: false
      required: [code, message, details, requestId]
      properties:
        code: { type: string, pattern: '^[A-Z][A-Z0-9_]+$' }
        message: { type: string, minLength: 1, maxLength: 500 }
        details: { type: [object, 'null'], additionalProperties: true }
        requestId: { type: string }
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [success, error]
      properties:
        success: { const: false }
        error: { $ref: '#/components/schemas/ErrorObject' }
...
    UpdateProfileRequest:
      type: object
      additionalProperties: false
      required: [expectedVersion]
      minProperties: 2
      properties:
        expectedVersion: { type: integer, minimum: 1 }
        username: { type: string, minLength: 3, maxLength: 30, pattern: '^[A-Za-z0-9_]+$' }
        displayName: { type: string, minLength: 1, maxLength: 80 }
        bio: { type: [string, 'null'], maxLength: 500 }
        countryCode: { type: string, pattern: '^[A-Z]{2}$' }
        preferredLanguage: { type: string, minLength: 2, maxLength: 16 }
        timezone: { type: string, minLength: 1, maxLength: 64 }
        dateOfBirth: { type: [string, 'null'], format: date }
        profileVisibility: { type: string, enum: [PUBLIC, FRIENDS, PRIVATE] }
    UpdatePreferencesRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        preferredGenreIds:
          { type: array, uniqueItems: true, maxItems: 50, items: { type: string, format: uuid } }
        dislikedGenreIds:
          { type: array, uniqueItems: true, maxItems: 50, items: { type: string, format: uuid } }
        preferredLanguages:
          { type: array, uniqueItems: true, maxItems: 30, items: { type: string, maxLength: 16 } }
        preferredCountries:
          {
            type: array,
            uniqueItems: true,
            maxItems: 30,
            items: { type: string, pattern: '^[A-Z]{2}$' },
          }
        preferredDecades:
          {
            type: array,
            uniqueItems: true,
            maxItems: 20,
            items: { type: integer, multipleOf: 10, minimum: 1880, maximum: 2200 },
          }
        preferredRuntimeMin: { type: [integer, 'null'], minimum: 1, maximum: 1440 }
        preferredRuntimeMax: { type: [integer, 'null'], minimum: 1, maximum: 1440 }
        preferredRatingSystem: { type: string, enum: [FIVE_STAR, TEN_POINT, LIKE_DISLIKE] }
        contentTypes:
          {
            type: array,
            uniqueItems: true,
            minItems: 1,
            items: { type: string, enum: [MOVIE, TV, ANIME, DOCUMENTARY, SHORT_FILM] },
          }
        spoilerPreference: { type: string, enum: [ALWAYS_HIDE, HIDE_UNTIL_REVEALED, SHOW] }
        adultContentEnabled: { type: boolean }
        theme: { type: string, enum: [SYSTEM, LIGHT, DARK] }
        defaultCountryForStreaming: { type: string, pattern: '^[A-Z]{2}$' }
        autoplayTrailers: { type: boolean }
        reduceMotion: { type: boolean }
        mainstreamPreferencePercent: { type: integer, minimum: 0, maximum: 100 }
    UpdatePrivacyRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        watchHistoryVisibility: { $ref: '#/components/schemas/Visibility' }
        ratingsVisibility: { $ref: '#/components/schemas/Visibility' }
        reviewsVisibility: { $ref: '#/components/schemas/Visibility' }
        listsVisibility: { $ref: '#/components/schemas/Visibility' }
        friendListVisibility: { $ref: '#/components/schemas/Visibility' }
        wrapsVisibility: { $ref: '#/components/schemas/Visibility' }
        onlineStatusVisibility: { $ref: '#/components/schemas/Visibility' }
        shareWatchActivity: { type: boolean }
        shareRatingActivity: { type: boolean }
        shareReviewActivity: { type: boolean }
        shareListActivity: { type: boolean }
        shareAchievementActivity: { type: boolean }
        leaderboardVisibility: { $ref: '#/components/schemas/Visibility' }
        passportVisibility: { $ref: '#/components/schemas/Visibility' }
    UpdateOnboardingRequest:
      type: object
      additionalProperties: false
      required: [step, data]
      properties:
        step:
          {
            type: string,
            enum:
              [
                PROFILE,
                CONTENT_TYPES,
                GENRES,
                FAVORITES,
                DISLIKES,
                STREAMING,
                RECOMMENDATIONS,
                SOCIAL,
                NOTIFICATIONS,
              ],
          }
        data: { type: object, additionalProperties: true }
    CompleteOnboardingRequest:
      type: object
      additionalProperties: false
      required: [acceptedPrivacyVersion, acceptedTermsVersion, expectedProfileVersion]
      properties:
        acceptedPrivacyVersion: { type: string, minLength: 1, maxLength: 40 }
        acceptedTermsVersion: { type: string, minLength: 1, maxLength: 40 }
        expectedProfileVersion: { type: integer, minimum: 1 }
...
    DeleteAccountRequest:
      type: object
      additionalProperties: false
      required: [confirmation, revokeImmediately]
      properties:
        confirmation: { type: string, const: DELETE MY ACCOUNT }
        revokeImmediately: { type: boolean, const: true }
        reasonCode: { type: string, maxLength: 80 }
...
    RecommendationFeedbackRequest:
      type: object
      additionalProperties: false
      required: [feedbackType]
      properties:
        feedbackType: { type: string, enum: [VIEWED, SAVED, DISMISSED, SELECTED] }
...